% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elbo_grad.R
\name{Elbo_grad}
\alias{Elbo_grad}
\title{Calculate the Elbo and the gradients}
\usage{
Elbo_grad(data, params)
}
\arguments{
\item{data}{list(Y, R, X)}

\item{params}{list(B, D, C, M, S)}
}
\value{
Elbo and gradients
}
\description{
Calculate the Elbo and the gradients
}
