% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{Init}
\alias{Init}
\title{Init}
\usage{
Init(Y, X, q)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}
}
\value{
A list with elements:
  \describe{
    \item{B}{Matrix of Poisson regression coefficients (d x p).}
    \item{C}{Matrix of latent structure estimates (p x q).}
    \item{M}{Matrix of variational parameters (n x p).}
    \item{S}{Matrix of variance parameters (n x q).}
  }
}
\description{
Parameters initialisation
}
