% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_ZIP_q0.R
\name{Init_ZIP_q0}
\alias{Init_ZIP_q0}
\title{Init_ZIP_q0}
\usage{
Init_ZIP_q0(Y, X, q)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}
}
\value{
A list with elements:
  \describe{
    \item{B}{Matrix of Poisson regression coefficients (d x p).}
    \item{D}{Matrix of logistic regression coefficients (d x p).}
  }
}
\description{
Parameters initialisation in the zero inflated case
}
