% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miss_PLNPCA.R
\name{Miss.PLNPCA}
\alias{Miss.PLNPCA}
\title{Miss.PLNPCA}
\usage{
Miss.PLNPCA(Y, X, O, w, q, params, config)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{O}{offsets}

\item{w}{weights}

\item{q}{size of the latent space}

\item{params}{Initial parameters}

\item{config}{configuration of the optimizer}
}
\value{
A list of the estimated parameters
}
\description{
Estimation of the parameters and the missing data
}
