% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miss_ZIPLNPCA_OneGradientDescent.R
\name{Miss.ZIPLNPCA_OneGradientDescent}
\alias{Miss.ZIPLNPCA_OneGradientDescent}
\title{Miss.ZIPLNPCA_OneGradientDescent}
\usage{
Miss.ZIPLNPCA_OneGradientDescent(
  Y,
  X,
  q,
  params,
  method = "Only_B",
  config = NULL,
  tolS = NULL
)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}

\item{params}{parameters including initial B}

\item{config}{configuration of the optimizer}

\item{tolS}{S tolerance}
}
\value{
A list of the estimated parameters
}
\description{
Estimation of the parameters and the missing data
}
