% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miss_ZIPLNPCA_Optim_S.R
\name{Miss.ZIPLNPCA_Optim.S}
\alias{Miss.ZIPLNPCA_Optim.S}
\title{Miss.ZIPLNPCA_Optim.S}
\usage{
Miss.ZIPLNPCA_Optim.S(
  Y,
  X,
  q,
  params = NULL,
  config = NULL,
  tolS = NULL,
  tolxi = NULL
)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}

\item{params}{Initial parameters}

\item{config}{configuration of the optimizer}

\item{tolS}{S tolerance}
}
\value{
A list of the estimated parameters
}
\description{
Estimation of the parameters and the missing data
}
