% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miss_ZIPLNPCA_VE.R
\name{Miss.ZIPLNPCA_VE}
\alias{Miss.ZIPLNPCA_VE}
\title{Miss.ZIPLNPCA_VE}
\usage{
Miss.ZIPLNPCA_VE(Y, X, q, params = NULL, config_vem = NULL, config = NULL)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}

\item{params}{Initial parameters}

\item{config_vem}{configuration of the optimizer}

\item{config}{configuration for the steps}
}
\value{
A list of the estimated parameters
}
\description{
Estimation of the parameters and the missing data
}
