% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\title{BIC}
\usage{
BIC(Y, X, fit, q)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{fit}{fitting}

\item{q}{size of the latent space}
}
\value{
A list with elements:
  \describe{
   \item{BIC} Bayesian information criterion
  }
}
\description{
Parameters initialisation
}
