% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{covmat}
\alias{covmat}
\title{Wrapper pour covmat du package lori}
\usage{
covmat(n, p, R = NULL, C = NULL, E = NULL, center = FALSE)
}
\arguments{
\item{n}{number of rows}

\item{p}{number ofcolumns}

\item{R}{nxK1 matrix of row covariates}

\item{C}{nxK2 matrix of column covariates}

\item{E}{(n+p)xK3 matrix of row-column covariates}

\item{center}{boolean indicating whether the returned covariate matrix should be centered (for identifiability)}
}
\value{
La matrice de covariables telle que calculée par \code{lori::covmat}.
}
\description{
Cette fonction appelle directement \code{lori::covmat}. 
Elle est fournie ici pour éviter à l'utilisateur de charger lori.
}
\examples{
if (requireNamespace("lori", quietly = TRUE)) {
  n <- 3; p <- 2
  R <- matrix(rnorm(n*2), n)
  C <- matrix(rnorm(p*2), p)
  X <- covmat(n, p, R, C)
  dim(X)  # 6 x 4
}
}
