% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elbo_grad_logS.R
\name{Elbo_grad_logS}
\alias{Elbo_grad_logS}
\title{Calculate the Elbo and the gradients}
\usage{
Elbo_grad_logS(data, params)
}
\arguments{
\item{data}{list(Y, R, X)}

\item{params}{list(B, C, M, S)}
}
\value{
Elbo and gradients
}
\description{
Calculate the Elbo and the gradients
}
