% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_sandwich.R
\name{IC}
\alias{IC}
\title{95% interval for a theta}
\usage{
IC(theta, var)
}
\arguments{
\item{theta}{vector theta.}

\item{var}{Associated variance (vector).}
}
\value{
Length-2 vector: lower and upper bounds.
}
\description{
Computes the 95% interval for theta.
}
