% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miss_ZIPLNPCA.R
\name{Miss.ZIPLNPCA}
\alias{Miss.ZIPLNPCA}
\title{Miss.ZIPLNPCA}
\usage{
Miss.ZIPLNPCA(Y, X, q, params = NULL, config = NULL, tolS = NULL)
}
\arguments{
\item{Y}{count matrix}

\item{X}{covariates}

\item{q}{size of the latent space}

\item{params}{Initial parameters}

\item{config}{configuration of the optimizer}

\item{tolS}{S tolerance}
}
\value{
A list of the estimated parameters
}
\description{
Estimation of the parameters and the missing data
}
